kill proc_holder pt9: - converts some / most mutations over

+++ code/datums/mutations/_mutations.dm
@@ -12,7 +12,7 @@
 	var/text_gain_indication = ""
 	var/text_lose_indication = ""
 	var/static/list/visual_indicators = list()
-	var/obj/effect/proc_holder/spell/power
+	var/datum/action/cooldown/spell/power_path
 	var/layer_used = MUTATIONS_LAYER //which mutation layer to use
 	var/list/species_allowed //to restrict mutation to only certain species
 	var/health_req //minimum health required to acquire the mutation
@@ -84,7 +84,7 @@
 		owner.remove_overlay(layer_used)
 		owner.overlays_standing[layer_used] = mut_overlay
 		owner.apply_overlay(layer_used)
-	grant_spell() //we do checks here so nothing about hulk getting magic
+	grant_power() //we do checks here so nothing about hulk getting magic
 	if(!modified)
 		addtimer(CALLBACK(src, .proc/modify, 5)) //gonna want children calling ..() to run first
 
@@ -108,8 +108,10 @@
 		mut_overlay.Remove(get_visual_indicator())
 		owner.overlays_standing[layer_used] = mut_overlay
 		owner.apply_overlay(layer_used)
-	if(power)
-		owner.RemoveSpell(power)
+	if(power_path)
+		// Any powers we made are linked to our mutation datum,
+		// so deleting ourself will also delete it and remove it
+		// ...Why don't all mutation delete on loss? Not sure
 		qdel(src)
 
 /mob/living/carbon/proc/update_mutations_overlay()
@@ -133,12 +135,21 @@
 				overlays_standing[CM.layer_used] = mut_overlay
 				apply_overlay(CM.layer_used)
 
-/datum/mutation/human/proc/modify() //called when a genome is applied so we can properly update some stats without having to remove and reapply the mutation from someone
-	if(modified || !power || !owner)
+/**
+ * Called when a chromosome is applied so we can properly update some stats
+ * without having to remove and reapply the mutation from someone
+ *
+ * Returns `null` if no modification was done, and
+ * returns an instance of a power if modification was complete
+ */
+/datum/mutation/human/proc/modify()
+	if(modified || !power_path || !owner)
 		return
-	power.charge_max *= GET_MUTATION_ENERGY(src)
-	power.charge_counter *= GET_MUTATION_ENERGY(src)
-	modified = TRUE
+	var/datum/action/cooldown/spell/modified_power = locate(power_path) in owner.actions
+	if(!modified_power)
+		CRASH("Genetic mutation [type] called modify(), but could not find a action to modify!")
+	modified_power.cooldown_time *= GET_MUTATION_ENERGY(src) // Doesn't do anything for mutations with energy_coeff unset
+	return modified_power
 
 /datum/mutation/human/proc/copy_mutation(datum/mutation/human/HM)
 	if(!HM)
@@ -167,14 +178,15 @@
 	else
 		qdel(src)
 
-/datum/mutation/human/proc/grant_spell()
-	if(!ispath(power) || !owner)
+/datum/mutation/human/proc/grant_power()
+	if(!ispath(power_path) || !owner)
 		return FALSE
 
-	power = new power()
-	power.action_background_icon_state = "bg_tech_blue_on"
-	power.panel = "Genetic"
-	owner.AddSpell(power)
+	var/datum/action/cooldown/spell/new_power = new power_path(src)
+	new_power.background_icon_state = "bg_tech_blue_on"
+	new_power.panel = "Genetic"
+	new_power.Grant(owner)
+
 	return TRUE
 
 // Runs through all the coefficients and uses this to determine which chromosomes the
